changed type of id_db_t into long long int (to correspond to BIGSERIAL)

+++ include/types.h
@@ -3,7 +3,7 @@
 
 typedef char url_t[255];
 typedef char ip_address_t[255];
-typedef int *id_db_t;
+typedef long long int *id_db_t;
 
 typedef enum
 {
